@@ -1,6 +1,6 @@
-import { test, expect } from "vitest";
 import { render } from "@testing-library/react";
-import App from "./App";
+import { expect, test } from "vitest";
+import { App } from "./App";
 
 test("renders learn react link", () => {
   const { getByText } = render(<App />);
