@@ -1,6 +1,6 @@
 import "./App.css";
 
-function App() {
+export const App = () => {
   return (
     <div className="App">
       <header className="App-header">
@@ -8,6 +8,4 @@ function App() {
       </header>
     </div>
   );
-}
-
-export default App;
+};
